TRAN-14: applied improvements after code review
