+++ transit/rest_api/urls.py
@@ -1,42 +1,15 @@
-from django.contrib.auth.models import User
-from django.urls import include, path
-from rest_framework import routers, serializers, viewsets
+from django.urls import (
+    include,
+    path
+)
+from rest_framework import routers
+from rest_framework_simplejwt.views import (
+    TokenRefreshView,
+    TokenVerifyView,
+    TokenObtainPairView
+)
 
-
-# Serializers define the API representation.
-from rest_framework.permissions import IsAuthenticated
-from rest_framework_simplejwt.views import TokenRefreshView, TokenVerifyView, TokenObtainPairView
-
-
-from transit.rest_api.permissions import IsUserManager
-
-
-class UserSerializer(serializers.HyperlinkedModelSerializer):
-    class Meta:
-        model = User
-        fields = ['url', 'username', 'email', 'is_staff', 'password']
-        extra_kwargs = {'password': {'write_only': True}}
-
-    def create(self, validated_data):
-        user = User(**validated_data)
-        user.set_password(validated_data['password'])
-        user.save()
-        return user
-
-
-class UserUpdateSerializer(serializers.HyperlinkedModelSerializer):
-    class Meta:
-        model = User
-        fields = ['url', 'username', 'email', 'is_staff']
-
-
-# ViewSets define the view behavior.
-class UserViewSet(viewsets.ModelViewSet):
-    permission_classes = (IsAuthenticated, IsUserManager)
-    queryset = User.objects.all()
-
-    def get_serializer_class(self):
-        return UserUpdateSerializer if self.request.method == 'PUT' else UserSerializer
+from transit.rest_api.users import UserViewSet
 
 
 # Routers provide an easy way of automatically determining the URL conf.
