+++ transit/rest_api/users/__init__.py
@@ -1,0 +1,5 @@
+from transit.rest_api.users.serializers import (
+    UserSerializer,
+    UserUpdateSerializer
+)
+from transit.rest_api.users.views import UserViewSet
