+++ transit/rest_api/users/serializers/__init__.py
@@ -1,0 +1,4 @@
+from transit.rest_api.users.serializers.user_serializer import (
+    UserSerializer,
+    UserUpdateSerializer
+)
