+++ transit/rest_api/users/views/__init__.py
@@ -1,0 +1,1 @@
+from transit.rest_api.users.views.user_view import UserViewSet
