+++ transit/tests/api_test/helpers/__init__.py
@@ -1,0 +1,1 @@
+from transit.tests.api_test.helpers.user_helper import UserTestHelper
