--- transit/tests/api_test/mixin/__init__.py
@@ -1,1 +1,0 @@
-from transit.tests.api_test.mixin.user_mixin import UserMixin
